requirements.txt
asyncpg==0.26.0
fastapi==0.79.0
gunicorn==20.1.0
inflect==6.0.0
orjson==3.7.12
python-dateutil==2.8.2
sqlalchemy==1.4.39
starlette_authlib==0.1.12
typing_extensions==4.3.0
uvicorn==0.18.2